pyproject.toml
[tool.poetry]
name = "dolipy"
version = "0.1.0"
description = "A simple wrapper for dolibarr api"
authors = ["Dimitrios Strantsalis <dstrants@gmail.com>"]

[tool.poetry.dependencies]
python = "3.8"
requests = "^2.25.1"
pydantic = "^1.7.3"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]
pylama = "^7.7.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
